remove old gotProblemSet and replace to have SSOT
